EDX-6482 Update to go 1.25 and golangci-lint 2.5.0
Also fixes all errors as reported by golangci-lint 2.5.0

+++ pkg/dtos/responses/filter.go
@@ -26,7 +26,7 @@ type MultiFiltersResponse struct {
 	Filters                              []dtos.Filter `json:"filters"`
 }
 
-func NewMultiFiltersResponse(requestId string, message string, statusCode int, totalCount uint32, filters []dtos.Filter) MultiFiltersResponse {
+func NewMultiFiltersResponse(requestId string, message string, statusCode int, totalCount int64, filters []dtos.Filter) MultiFiltersResponse {
 	return MultiFiltersResponse{
 		BaseWithTotalCountResponse: dtoCommon.NewBaseWithTotalCountResponse(requestId, message, statusCode, totalCount),
 		Filters:                    filters,
